requirements.txt
# GUI
customtkinter>=5.2.0
pillow>=10.0.0

# Audio
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.11.0

# Groq API
groq>=0.4.0

# Hotkeys and clipboard
keyboard>=0.13.5
pyperclip>=1.8.2
pyautogui>=0.9.54

# Environment and config
python-dotenv>=1.0.0

# For building installer
pyinstaller>=6.0.0
